+++ packages/windie-sdk-js/cjs/transport/WindieAgentSession.js
@@ -17,4 +17,9 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
     for (var p in m) if (p !== "default" && !Object.prototype.hasOwnProperty.call(exports, p)) __createBinding(exports, m, p);
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.createWindieAgentSession = exports.createWindieAgentBackendTransport = exports.WindieAgentSession = void 0;
+const AgentSession_js_1 = require("./AgentSession.js");
 __exportStar(require("./AgentSession.js"), exports);
+Object.defineProperty(exports, "WindieAgentSession", { enumerable: true, get: function () { return AgentSession_js_1.AgentSession; } });
+Object.defineProperty(exports, "createWindieAgentBackendTransport", { enumerable: true, get: function () { return AgentSession_js_1.createAgentBackendTransport; } });
+Object.defineProperty(exports, "createWindieAgentSession", { enumerable: true, get: function () { return AgentSession_js_1.createAgentSession; } });